[DAD-494] feat: 스크랩 템플릿 컴포넌트 분리
