+++ src/components/templates/ListTemplate.tsx
@@ -1,83 +1,14 @@
-import styled from 'styled-components';
+import ExistOtherScrapContainer from '../organisms/ExistOtherScrapContainer';
+import ScrapListHeader from '../molcules/ScrapListHeader';
+import EmptyScrapContainer from '../organisms/EmptyScrapContainer';
 
-import theme from '../../assets/styles/theme';
-import searchIcon from '../../assets/icons/SearchIcon.png';
-
-function OtherTemplate() {
-
+function ListTemplate({ lists }) {
     return (
-        <ScrapListContainer>
-            <ScrapListHeader>
-                <ScarpCountWrapper>
-                    <EmpasizedTypography>0 </EmpasizedTypography>
-                    <DefaultTypography>개의 </DefaultTypography>
-                    <EmpasizedTypography>전체 스크랩</EmpasizedTypography>
-                    <DefaultTypography>이 있습니다.</DefaultTypography>
-                </ScarpCountWrapper>
-                <SearchBar>
-                    <SearchIconWrapper src={searchIcon} />
-                    <EmpasizedTypography>Search</EmpasizedTypography>
-                </SearchBar>
-            </ScrapListHeader>
-            <ScrapList>
-                <EmpasizedTypography>스크랩을 추가해주세요</EmpasizedTypography>
-            </ScrapList>
-        </ScrapListContainer>
+        <>
+            <ScrapListHeader type='전체' count={lists.length} />
+            {lists.length ? <ExistOtherScrapContainer contents={lists} /> : <EmptyScrapContainer />}
+        </>
     )
 }
 
-const ScrapListContainer = styled.div`
-    width: calc(100% - 200px);
-    height: 100%;
-    background-color: ${theme.color.background_color};
-    position: fixed;
-    right: 0;
-    top: 50px; 
-    @media screen and (max-width: 600px) {
-      width: 100vw;
-      left: 0;
-    }
-    display: flex;
-    flex-direction: column;
-`
-
-const ScrapListHeader = styled.div`
-    display: flex;
-    padding: 20px;
-    justify-content: space-between;
-`
-
-const ScrapList = styled.div`
-    display: flex;
-    justify-content: center;
-    flex: 1;
-    align-items: center;
-`
-
-const ScarpCountWrapper = styled.div`
-
-`
-
-const DefaultTypography = styled.span`
-    font-size: 14px;
-    color: ${theme.color.text_gray_color};
-`
-
-const EmpasizedTypography = styled.span`
-    font-size: 14px;
-    font-weight: bold;
-    color: ${theme.color.text_gray_color};
-`
-
-const SearchBar = styled.div`
-    display: flex;
-    align-items: center;
-    gap: 10px;
-`
-
-const SearchIconWrapper = styled.img`
-    width: 24px;
-    height: 24px;
-`
-
-export default OtherTemplate;
+export default ListTemplate;
